Alteraçoes gerais e correçoes de bugs. Imagens agora nao substituem as antigas, simplismente é criado uma nova e redirecionado o LogoId. Cards do dashboard agora sao azuis acima de 120%.

+++ Web/Gamific.Web/Controllers/Public/DashboardController.cs
@@ -87,7 +87,7 @@ namespace Vlast.Gamific.Web.Controllers.Public
 
                 if (results[0] != null) { 
                     point.Add(i);
-                    point.Add(results[0].TotalPoints);
+                    point.Add((int)results[0].TotalPoints);
                     chartDTO.Positions.Add(point);
                     i++;
                 }
